+++ example/lib/app/app_module.dart
@@ -1,3 +1,5 @@
+import 'package:example/app/shared/blocs/auth_bloc.dart';
+import 'package:example/app/shared/services/service_service.dart';
 import 'package:example/app/app_bloc.dart';
 import 'package:bloc_pattern/bloc_pattern.dart';
 import 'package:flutter/material.dart';
@@ -6,11 +8,14 @@ import 'package:example/app/app_widget.dart';
 class AppModule extends ModuleWidget {
   @override
   List<Bloc> get blocs => [
+        Bloc((i) => AuthBloc()),
         Bloc((i) => AppBloc()),
       ];
 
   @override
-  List<Dependency> get dependencies => [];
+  List<Dependency> get dependencies => [
+        Dependency((i) => ServiceService()),
+      ];
 
   @override
   Widget get view => AppWidget();